requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
packaging==24.2
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
